Split up filestore into 3
